requirements.txt
asgiref==3.4.1
Babel==2.10.3
dbr==9.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==21.1
django-import-export==2.7.1
django-mathfilters==1.0.0
django-money==2.1.1
django-six==1.0.4
django-widget-tweaks==1.4.11
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.1
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.10
pandas==0.25.3
Pillow==8.4.0
py-moneyed==1.2
python-barcode==0.14.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzbar==0.1.9
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.1.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
